Cadmus.Seed.Img.Parts.Test/GalleryImageAnnotationsPartSeederTest.cs
using Cadmus.Core;
using Cadmus.Img.Parts;
using Fusi.Tools.Configuration;
using System;
using System.Reflection;
using Xunit;

namespace Cadmus.Seed.Img.Parts.Test;

public sealed class GalleryImageAnnotationsPartSeederTest
{
    private static readonly PartSeederFactory _factory = TestHelper.GetFactory();
    private static readonly SeedOptions _seedOptions = _factory.GetSeedOptions();
    private static readonly IItem _item =
        _factory.GetItemSeeder().GetItem(1, "facet");

    [Fact]
    public void TypeHasTagAttribute()
    {
        Type t = typeof(GalleryImageAnnotationsPartSeeder);
        TagAttribute? attr = t.GetTypeInfo().GetCustomAttribute<TagAttribute>();
        Assert.NotNull(attr);
        Assert.Equal("seed.it.vedph.img.gallery-image-annotations", attr!.Tag);
    }

    [Fact]
    public void Seed_Ok()
    {
        GalleryImageAnnotationsPartSeeder seeder = new();
        seeder.SetSeedOptions(_seedOptions);

        IPart? part = seeder.GetPart(_item, null, _factory);

        Assert.NotNull(part);

        GalleryImageAnnotationsPart? p = part as GalleryImageAnnotationsPart;
        Assert.NotNull(p);

        TestHelper.AssertPartMetadata(p!);

        Assert.NotEmpty(p!.Annotations);
    }
}
